ported to libc++ stdlib. compiles until the fuzzer target. needed to recompile boost locally.
1. Environment setup

     export MONAD_ROOT=/home/abhishek/monad-latest/monad
     export BOOST_PREFIX=$HOME/boost-libcxx
     export LIBBACKTRACE_ROOT=$HOME/libbacktrace-libcxx
  2. Build libbacktrace (static, libc++)

     cd $HOME
     git clone https://github.com/ianlancetaylor/libbacktrace.git
     cd libbacktrace
     ./configure CC=clang-19 CFLAGS='-O2 -fPIC' --prefix=$LIBBACKTRACE_ROOT
     make -j$(nproc)
     make install
  3. Build Boost 1.83.0 against libc++

     cd $HOME
     curl -LO https://archives.boost.io/release/1.83.0/source/boost_1_83_0.tar.bz2
     tar xf boost_1_83_0.tar.bz2
     cd boost_1_83_0
     ./bootstrap.sh --with-toolset=clang --prefix=$BOOST_PREFIX

     ./b2 toolset=clang-19 \
          cxxflags="-stdlib=libc++ -march=haswell -fPIC -I$LIBBACKTRACE_ROOT/include" \
          linkflags="-stdlib=libc++ -L$LIBBACKTRACE_ROOT/lib" \
          --with-context --with-fiber --with-stacktrace \
          --with-filesystem --with-json --with-container --with-atomic --with-thread \
          -sLIBBACKTRACE_PATH=$LIBBACKTRACE_ROOT \
          install -j$(nproc)
  4. Reconfigure the project for libc++

     cd $MONAD_ROOT
     python3 - <<'PY'

  import shutil, os
  shutil.rmtree('build', ignore_errors=True)
  os.mkdir('build')
  PY

  cmake -S . -B build -G Ninja \
  -DCMAKE_C_COMPILER=/usr/bin/clang-19 \
  -DCMAKE_CXX_COMPILER=/usr/bin/clang++-19 \
  -DCMAKE_C_FLAGS='-march=haswell' \
  -DCMAKE_CXX_FLAGS='-march=haswell -stdlib=libc++' \
  -DCMAKE_PREFIX_PATH=$BOOST_PREFIX \
  -DLIBBACKTRACE_ROOT=$LIBBACKTRACE_ROOT \
  -Dboost_context_DIR=$BOOST_PREFIX/lib/cmake/boost_context-1.83.0 \
  -Dboost_fiber_DIR=$BOOST_PREFIX/lib/cmake/boost_fiber-1.83.0 \
  -Dboost_stacktrace_backtrace_DIR=$BOOST_PREFIX/lib/cmake/boost_stacktrace_backtrace-1.83.0 \
  -Dboost_stacktrace_addr2line_DIR=$BOOST_PREFIX/lib/cmake/boost_stacktrace_addr2line-1.83.0 \
  -Dboost_stacktrace_basic_DIR=$BOOST_PREFIX/lib/cmake/boost_stacktrace_basic-1.83.0 \
  -Dboost_filesystem_DIR=$BOOST_PREFIX/lib/cmake/boost_filesystem-1.83.0 \
  -Dboost_json_DIR=$BOOST_PREFIX/lib/cmake/boost_json-1.83.0 \
  -Dboost_container_DIR=$BOOST_PREFIX/lib/cmake/boost_container-1.83.0 \
  -Dboost_atomic_DIR=$BOOST_PREFIX/lib/cmake/boost_atomic-1.83.0 \
  -Dboost_thread_DIR=$BOOST_PREFIX/lib/cmake/boost_thread-1.83.0 \
  -DBoost_USE_STATIC_LIBS=ON \
  -DBoost_USE_STATIC_RUNTIME=OFF

  5. **Build**
  ```bash
  ./scripts/build.sh   # (or plain `ninja -C build`)

  That reproduces exactly the libc++ toolchain we just set up (Boost + libbacktrace in $HOME, Clang 19, -stdlib=libc++).

+++ category/mpt/update_aux.cpp
@@ -1141,12 +1141,12 @@ Node::SharedPtr UpdateAuxImpl::do_update(
         physical_to_virtual(node_writer_slow->sender().offset());
     LOG_INFO_CFORMAT(
         "Finish upserting version %lu. Min valid version %lu. Time elapsed: "
-        "%ld us. Disk usage: %.4f. Chunks: %u fast, %u slow, %u free. Writer "
+        "%lld us. Disk usage: %.4f. Chunks: %u fast, %u slow, %u free. Writer "
         "offsets: fast={%u,%u}, slow={%u,%u}. Compaction head offset fast=%u, "
         "slow=%u",
         version,
         db_history_min_valid_version(),
-        duration.count(),
+        static_cast<long long>(duration.count()),
         disk_usage(),
         num_chunks(chunk_list::fast),
         num_chunks(chunk_list::slow),
@@ -1159,9 +1159,9 @@ Node::SharedPtr UpdateAuxImpl::do_update(
         (uint32_t)compact_offset_slow);
     if (duration > std::chrono::microseconds(500'000)) {
         LOG_WARNING_CFORMAT(
-            "Upsert version %lu takes longer than 0.5 s, time elapsed: %ld us.",
+            "Upsert version %lu takes longer than 0.5 s, time elapsed: %lld us.",
             version,
-            duration.count());
+            static_cast<long long>(duration.count()));
     }
     return root;
 }
